+++ objc2-foundation/src/string.rs
@@ -2,6 +2,8 @@ use alloc::borrow::ToOwned;
 use core::cmp;
 use core::ffi::c_void;
 use core::fmt;
+use core::panic::RefUnwindSafe;
+use core::panic::UnwindSafe;
 use core::ptr::NonNull;
 use core::slice;
 use core::str;
@@ -41,6 +43,9 @@ object! {
 // TODO: SAFETY
 unsafe impl Sync for NSString {}
 unsafe impl Send for NSString {}
+
+impl UnwindSafe for NSString {}
+impl RefUnwindSafe for NSString {}
 
 impl NSString {
     unsafe_def_fn! {
@@ -121,6 +126,8 @@ impl NSString {
     /// TODO: Further explain this.
     #[doc(alias = "UTF8String")]
     pub fn as_str<'r, 's: 'r, 'p: 'r>(&'s self, pool: &'p AutoreleasePool) -> &'r str {
+        // NOTE: Please keep up to date with `objc2::exception`!
+
         // This is necessary until `auto` types stabilizes.
         pool.__verify_is_inner();
 
@@ -155,6 +162,8 @@ impl NSString {
 
         // TODO: Always UTF-8, so should we use `from_utf8_unchecked`?
         str::from_utf8(bytes).unwrap()
+
+        // NOTE: Please keep up to date with `objc2::exception`!
     }
 
     // TODO: Allow usecases where the NUL byte from `UTF8String` is kept?